teams module completed!

+++ database/migrations/2022_06_09_173106_create_teams_table.php
@@ -19,7 +19,7 @@ class CreateTeamsTable extends Migration
             $table->string('name')->nullable();
             $table->bigInteger('sports_id');
             $table->bigInteger('leagues_id');
-            $table->decimal('points')->default(0);
+            $table->integer('points')->default(0);
             $table->timestamps();
         });
     }
